refac: fix the type of `K`-like to 32 bit

+++ tachyon/math/finite_fields/prime_field_base.h
@@ -117,8 +117,7 @@ class PrimeFieldBase : public FiniteField<F> {
         omega.SquareInPlace();
       }
     } else {
-      uint64_t log_size_of_group =
-          static_cast<uint64_t>(base::bits::Log2Ceiling(n));
+      uint32_t log_size_of_group = base::bits::Log2Ceiling(n);
       uint64_t size = uint64_t{1} << log_size_of_group;
 
       if (n != size || log_size_of_group > Config::kTwoAdicity) {
@@ -126,7 +125,7 @@ class PrimeFieldBase : public FiniteField<F> {
       }
 
       omega = F::FromMontgomery(Config::kTwoAdicRootOfUnity);
-      for (size_t i = log_size_of_group; i < Config::kTwoAdicity; ++i) {
+      for (uint32_t i = log_size_of_group; i < Config::kTwoAdicity; ++i) {
         omega.SquareInPlace();
       }
     }
